优化分类管理

+++ src/main/java/com/sht/deal/service/Classify2Service.java
@@ -1,11 +1,15 @@
 package com.sht.deal.service;
 
 import com.sht.deal.Mapper.Classify2Mapper;
+import com.sht.deal.domain.Classify1;
 import com.sht.deal.domain.Classify2;
+import com.sht.deal.exception.AllException;
 import com.sht.deal.service.GoodsService;
 
 import java.util.List;
+import java.util.Objects;
 
+import com.sht.deal.utils.JsonData;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,24 +24,40 @@ public class Classify2Service {
     @Autowired
     private UploadService uploadService;
 
-    public int add(Classify2 classify2) {
-        return this.classify2Mapper.insertSelective(classify2);
+    public JsonData add(Classify2 classify2) {
+        int i = this.classify2Mapper.insertSelective(classify2);
+        if (i != 1){
+            return JsonData.buildError("失败");
+        }
+        return JsonData.buildSuccess("成功");
     }
 
 
-    public int delete(int id) {
+    public JsonData delete(int id) {
         Classify2 classify2 = classify2Mapper.selectByPrimaryKey(id);
-        this.uploadService.deleteImage(classify2.getImage());
-        return this.classify2Mapper.deleteByPrimaryKey(id);
+        try {
+            this.classify2Mapper.deleteByPrimaryKey(id);
+            this.uploadService.deleteImage(classify2.getImage());
+            return JsonData.buildSuccess("删除成功");
+        }catch (Exception e){
+            return JsonData.buildError("删除失败,分类: " + classify2.getName() +",尚关联的有商品");
+        }
+
     }
 
-    public int update(Classify2 classify2) {
+    public JsonData update(Classify2 classify2) {
+
         if (classify2.getImage() != null) {
             Classify2 classify = classify2Mapper.selectByPrimaryKey(classify2.getId());
-            this.uploadService.deleteImage(classify.getImage());
-            return this.classify2Mapper.updateByPrimaryKeySelective(classify2);
+            if (!classify2.getImage().equals(classify.getImage())){
+                this.uploadService.deleteImage(classify.getImage());
+            }
         }
-        return this.classify2Mapper.updateByPrimaryKeySelective(classify2);
+        int i = this.classify2Mapper.updateByPrimaryKeySelective(classify2);
+        if (i != 1){
+            return JsonData.buildError("更新失败");
+        }
+        return JsonData.buildSuccess("更新成功");
     }
